Moved match helpers into their own file
